refactor(store-core): extract user-question handlers into handlers/user-question.ts (audit P2-3)
Move the user-question + user-input family (handleUserQuestion, handleUserInput,
OTHER_OPTION_VALUE/OTHER_OPTION_LABEL + their payload interfaces) out of the
handlers/index.ts god-module into handlers/user-question.ts, behind the barrel's
export * re-export so the public surface is unchanged.

Drop the now-unused parseUserInputMessage import (from ../user-input-handler)
from index.ts — it was only used by handleUserInput. ChatMessage and
nextMessageId stay (still used by permission handlers).

Pure move — body byte-identical to origin/main. index.ts 1557 -> 1335 lines.
